:truck: Move main file to src
