+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@wolfox/caravel",
   "version": "0.1.0",
-  "main": "main.js",
+  "main": "src/main.js",
   "description": "Smart, simple and efficient way to manage PostgreSQL migration files.",
   "repository": "git@github.com:wolfoxco/caravel.git",
   "homepage": "https://github.com/wofoxco/caravel",
